+++ projects/core/src/main/kotlin/site/siredvin/cloudsolutions/subsystems/Management.kt
@@ -2,16 +2,50 @@ package site.siredvin.cloudsolutions.subsystems
 
 import net.minecraft.server.MinecraftServer
 import site.siredvin.cloudsolutions.common.configuration.ModConfig
+import site.siredvin.cloudsolutions.computercraft.peripheral.KVStoragePeripheral
 import site.siredvin.cloudsolutions.subsystems.kv.DisabledKVManager
 import site.siredvin.cloudsolutions.subsystems.kv.KeyValueManager
 import site.siredvin.cloudsolutions.subsystems.kv.sqlite.KVSQLiteManager
 import site.siredvin.cloudsolutions.subsystems.statsq.StatsDClient
+import java.util.Collections
+import java.util.WeakHashMap
 import java.util.concurrent.Executors
 import java.util.concurrent.ScheduledExecutorService
 
 enum class KVStorageMode {
     DISABLED,
     SQLITE,
+}
+
+object SubscriptionManager {
+    private val kvStorages: MutableMap<String, MutableSet<KVStoragePeripheral>> = mutableMapOf()
+
+    fun addKVStorage(ownerUUID: String, kvStorage: KVStoragePeripheral) {
+        if (!kvStorages.contains(ownerUUID)) {
+            kvStorages[ownerUUID] = Collections.newSetFromMap(WeakHashMap())
+        }
+        kvStorages[ownerUUID]!!.add(kvStorage)
+    }
+
+    fun removeKVStorage(ownerUUID: String, kvStorage: KVStoragePeripheral) {
+        kvStorages[ownerUUID]?.remove(kvStorage)
+    }
+
+    fun onKeyChanged(ownerUUID: String, key: String, value: String) {
+        if (kvStorages.contains(ownerUUID)) {
+            kvStorages[ownerUUID]!!.forEach {
+                it.onKeyChanged(key, value)
+            }
+        }
+    }
+
+    fun onKeyDeleted(ownerUUID: String, key: String) {
+        if (kvStorages.contains(ownerUUID)) {
+            kvStorages[ownerUUID]!!.forEach {
+                it.onKeyDeleted(key)
+            }
+        }
+    }
 }
 
 object SubsystemManager {
@@ -24,6 +58,8 @@ object SubsystemManager {
             KVStorageMode.SQLITE -> KVSQLiteManager
         }
         kvManager?.init(server, executorService)
+        kvManager?.setOnKeyDeletedHook(SubscriptionManager::onKeyDeleted)
+        kvManager?.setOnKeyChangedHook(SubscriptionManager::onKeyChanged)
         if (ModConfig.enableStatsDConnection) {
             StatsDClient.init()
         }