+++ projects/core/src/main/kotlin/site/siredvin/cloudsolutions/subsystems/kv/KVKeyChangedHook.kt
@@ -1,0 +1,7 @@
+package site.siredvin.cloudsolutions.subsystems.kv
+
+import java.util.UUID
+
+fun interface KVKeyChangedHook {
+    fun handle(ownerUUID: String, key: String, value: String)
+}