+++ projects/core/src/main/kotlin/site/siredvin/cloudsolutions/subsystems/kv/KVKeyDeletedHook.kt
@@ -1,0 +1,5 @@
+package site.siredvin.cloudsolutions.subsystems.kv
+
+fun interface KVKeyDeletedHook {
+    fun handle(ownerUUID: String, key: String)
+}